refactor(backdrop): Rename backdrop module to ui
It'll serve as a module for all common components used in the catalog app
