+++ catalog/src/main/kotlin/com/adevinta/spark/catalog/CatalogApp.kt
@@ -63,10 +63,6 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 import com.adevinta.spark.SparkFeatureFlag
 import com.adevinta.spark.SparkTheme
-import com.adevinta.spark.catalog.backdrop.BackdropScaffold
-import com.adevinta.spark.catalog.backdrop.BackdropScaffoldDefaults
-import com.adevinta.spark.catalog.backdrop.BackdropValue
-import com.adevinta.spark.catalog.backdrop.rememberBackdropScaffoldState
 import com.adevinta.spark.catalog.configurator.ConfiguratorComponentsScreen
 import com.adevinta.spark.catalog.examples.ComponentsScreen
 import com.adevinta.spark.catalog.icons.IconDemoScreen
@@ -87,6 +83,10 @@ import com.adevinta.spark.catalog.themes.themeprovider.kleinanzeigen.Kleinanzeig
 import com.adevinta.spark.catalog.themes.themeprovider.leboncoin.LeboncoinTheme
 import com.adevinta.spark.catalog.themes.themeprovider.milanuncios.MilanunciosTheme
 import com.adevinta.spark.catalog.themes.themeprovider.subito.SubitoTheme
+import com.adevinta.spark.catalog.ui.BackdropScaffold
+import com.adevinta.spark.catalog.ui.BackdropScaffoldDefaults
+import com.adevinta.spark.catalog.ui.BackdropValue
+import com.adevinta.spark.catalog.ui.rememberBackdropScaffoldState
 import com.airbnb.android.showkase.models.ShowkaseBrowserComponent
 import com.google.accompanist.testharness.TestHarness
 import kotlinx.coroutines.launch
